コンポへの切り出し、不要な変数の削除を行いました。
